aakeedo_kanji/pyproject.toml
[tool.poetry]
name = "arcade_aakeedo_kanji"
version = "0.0.1"
description = "AI toolkit to help learn kanji"
authors = ["jottakka <fjlimal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "^1.0.5"
pytz = "^2025.2"
python-dateutil = "^2.9.0.post0"
pydantic-settings = "^2.9.1"
pytest-mock = "^3.14.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"
pytest-cov = "^4.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["arcade_aakeedo_kanji/**/*.py"]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true
